requirements.txt
django
pillow
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter